requirements.txt
djangorestframework==3.9.2
django==1.11
amqp==2.4.2
redis==3.2.1
celery==4.3.0
pascal_voc_writer==0.1.4
tensorflow==1.13.1
google-cloud-storage==1.14.0
opencv-python==4.1.0.25
tensorflow-hub
opencv-contrib-python-headless
google-cloud-vision==0.36.0
google-cloud-datastore==1.7.3
gunicorn==19.9.0
google-api-python-client==1.7.9
pathlib==1.0.1
oauth2client==4.1.3

absl-py==0.6.1
altgraph==0.10.2
appnope==0.1.0
argh==0.26.2
astor==0.7.1
astroid==1.6.5
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bdist-mpkg==0.5.0
bleach==3.0.2
configparser==3.5.0
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
enum34==1.1.6
evernote==1.25.3
fire==0.1.3
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
gast==0.2.1
grpcio==1.17.1
h5py==2.9.0
httplib2==0.12.0
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
macholib==1.5.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
modulegraph==0.10.4
nbconvert==5.4.0
nbformat==4.4.0
nose==1.3.7
notebook==5.7.2
numpy==1.15.4
oauth2==1.9.0.post1
pandocfilters==1.4.2
pathlib2==2.3.3
pathtools==0.1.2
pbr==5.1.1
pdf2image==1.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.4.2
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
py2app==0.7.3
Pygments==2.3.0
pylint==1.9.3
subprocess32==3.5.3
tensorboard==1.13.0

termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
tqdm==4.29.0
traitlets==4.3.2
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
xattr==0.6.4
xlrd==1.1.0